Fixed deleteOldData bug

+++ app/src/main/java/me/kirimin/mitsumine/data/TopData.kt
@@ -25,7 +25,7 @@ class TopData {
         KeywordDAO.delete(keyword)
     }
 
-    fun deleteOldFeedData(days: Long) {
+    fun deleteOldFeedData(days: Int) {
         FeedDAO.deleteOldData(days)
     }
 }
